fix func arg

+++ src/utils.h
@@ -1,6 +1,7 @@
 #ifndef UTILS_HPP
 #define UTILS_HPP
 
+#include <inttypes.h>
 #include <stdbool.h>
 #include <time.h>
 
@@ -24,7 +25,7 @@ extern int hms_to_seconds(int hour, int min, int sec);
 extern int hms_to_seconds_str(const char* str);
 extern int now_seconds(void);
 extern size_t total_seconds_to_string(char* buffer, size_t buffer_size, long total_seconds, bool append_total_seconds);
-extern void human_readable_time(char* buffer, size_t buffer_size, size_t seconds);
+extern void human_readable_time(char* buffer, size_t buffer_size, uint64_t seconds);
 extern void trim_spaces(char* buffer);
 extern void trim_left(char* buffer, size_t n);
 extern void trim_right(char* buffer, size_t n);
